Moved DeviceTrackingPoint methods to TrackingData and refractored code
